temp commit

+++ src/main/java/com/techstackgo/ecommerce/model/Cart.java
@@ -15,17 +15,17 @@ public class Cart {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     @OneToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name="user_id",nullable = false)
+    @JoinColumn(name = "user_id", nullable = false)
     private User user;
-    @OneToMany(mappedBy = "cart",cascade = CascadeType.ALL,orphanRemoval = true)
-    @Column(name="cart_items")
+    @OneToMany(mappedBy = "cart", cascade = CascadeType.ALL, orphanRemoval = true)
+    @Column(name = "cart_items")
     private Set<CartItem> cartItems = new HashSet<>();
-    @Column(name="total_price")
+    @Column(name = "total_price")
     private double totalPrice;
-    @Column(name="total_item")
-    private double totalItem;
-    @Column(name="total_discount")
+    @Column(name = "total_item")
+    private int totalItem;
+    @Column(name = "total_discount")
     private double totalDiscount;
-    @Column(name="final_price")
+    @Column(name = "final_price")
     private double finalPrice;
 }
